requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.1.7
django-autocomplete-light==3.12.1
django-crispy-forms==2.3
django-debug-toolbar==5.1.0
django-extensions==3.2.3
django-widget-tweaks==1.5.0
Faker==37.1.0
pillow==11.1.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2